fix: pass the prediction stride as ptrdiff_t
The intra predictors declared it as int while both the x86 and the NEON
assembly read the whole register: vp8_intrapred.asm does sub r0, r1 and
[r0+r1*2] without a movsxd, vp8pred_neon.S uses x1 and never w1. Neither
ABI defines the upper half of a register holding a 32-bit argument, so
the assembly was addressing off bits the caller had no duty to set. The
assembly is FFmpeg's, where the stride has always been a ptrdiff_t; only
the C prototypes were narrowed in the port.

Widening them is also what checkasm already assumed, so this clears the
three -fsanitize=function reports that made checkasm abort under
scripts/sanitize.sh.

+++ src/vp8pred.h
@@ -30,9 +30,9 @@ enum VP8Pred8x8Mode {
 
 typedef struct VP8PredContext {
     void (*pred4x4[VP8_PRED4X4_COUNT])(uint8_t *src, const uint8_t *topright,
-                                       int stride);
-    void (*pred8x8[VP8_PRED8X8_COUNT])(uint8_t *src, int stride);
-    void (*pred16x16[VP8_PRED8X8_COUNT])(uint8_t *src, int stride);
+                                       ptrdiff_t stride);
+    void (*pred8x8[VP8_PRED8X8_COUNT])(uint8_t *src, ptrdiff_t stride);
+    void (*pred16x16[VP8_PRED8X8_COUNT])(uint8_t *src, ptrdiff_t stride);
 } VP8PredContext;
 
 void ff_vp8_pred_init(VP8PredContext *pred);
